Move DataLoaderRegistryBuilder into dataloader package
